Without 7 and 8.

+++ StringCalculator-master/StringCalculator.Core/Interfaces/IStringCalculator.cs
@@ -2,6 +2,6 @@
 {
     public interface IStringCalculator
     {
-        int Add(string numbersString);
+        long Add(string numbersString);
     }
 }
